pyproject.toml

[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.17.2",
    "beautifulsoup4>=4.14.3",
    "chromadb==1.4.0",
    "fastapi==0.127.1",
    "langchain==1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-ollama==1.0.1",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-text-splitters>=1.1.0",
    "langfuse==3.11.2",
    "langgraph==1.0.5",
    "langgraph-checkpoint-postgres>=3.0.2",
    "litellm>=1.80.11",
    "psycopg[binary]==3.3.2",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "pymupdf>=1.26.7",
    "python-docx>=1.2.0",
    "python-multipart==0.0.21",
    "pytest==8.3.4",
    "ragas==0.4.2",
    "sqlalchemy==2.0.45",
    "streamlit>=1.52.2",
    "streamlit-extras>=0.7.8",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "torchvision==0.23.0",
    "uvicorn>=0.40.0",
]


[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
torchaudio = { index = "pytorch-cu128" }


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_rag"]
